[#24] unify logging to console
refacto every message to the console to use jirac_log for consistency
add a log-level option to set a level of details for JIRAC logs to the console : 0 = ERROR, 1 = INFO, 2 = DEBUG
option silent-mode is now implemented as setting the log level to -1 (ie. no log at all)
remove JIRAC header and victory ASCII arts because they are no logs and can not be controlled with log-level
refacto all program exits paths to use the same functions
move exit functions to core_functions.sh for consistency (ie. with the other log and exit related functions)